# backend/internal/http/docs/openapi.yaml
openapi: 3.0.3
info:
  title: HRIS LDP API
  version: 1.0.0
  description: Dokumentasi endpoint utama HRIS LDP.
servers:
  - url: /api
    description: API endpoints
  - url: /
    description: Public/root endpoints
...
paths:
...
  /super-admin/accounts:
    get:
      tags: [SuperAdmin]
      summary: Ambil data /super-admin/accounts
      security:
        - cookieAuth: []
      responses:
        '200':
          description: OK
        '401':
          description: Unauthenticated
    post:
      tags: [SuperAdmin]
      summary: Proses request /super-admin/accounts
      security:
        - cookieAuth: []
      responses:
        '200':
          description: OK
        '401':
          description: Unauthenticated
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: hris_session
...